--- src/main/java/org/javamoney/moneta/convert/ExchangeCurrencyOperator.java
@@ -1,38 +1,0 @@
-/**
- * Copyright (c) 2012, 2015, Credit Suisse (Anatole Tresch), Werner Keil and others by the @author tag.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package org.javamoney.moneta.convert;
-
-import java.util.Objects;
-
-import javax.money.CurrencyUnit;
-import javax.money.MonetaryAmount;
-import javax.money.MonetaryOperator;
-
-public class ExchangeCurrencyOperator implements MonetaryOperator {
-
-	private final CurrencyUnit currency;
-
-	ExchangeCurrencyOperator(CurrencyUnit currency) {
-		this.currency = currency;
-	}
-
-	@Override
-	public MonetaryAmount apply(MonetaryAmount amount) {
-		Objects.requireNonNull(amount, "Amount required.");
-		return amount.getFactory().setCurrency(currency).create();
-	}
-
-}
