+++ src/main/resources/java-money/defaults/IMF/rms_five.xls
