+++ src/test/java/org/javamoney/moneta/function/ExchangeRateMonetarySummaryStatisticsTest.java
@@ -1,4 +1,4 @@
-package org.javamoney.moneta.convert;
+package org.javamoney.moneta.function;
 
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertNotSame;
@@ -11,7 +11,6 @@ import javax.money.convert.ExchangeRateProvider;
 import javax.money.convert.MonetaryConversions;
 
 import org.javamoney.moneta.Money;
-import org.javamoney.moneta.function.MonetarySummaryStatistics;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
@@ -70,7 +69,7 @@ public class ExchangeRateMonetarySummaryStatisticsTest {
 	@Test
 	public void combineImplSummaryTest() {
 		MonetarySummaryStatistics summaryA = createSummary(BRAZILIAN_REAL);
-		MonetarySummaryStatistics summaryB = createSummary(DOLLAR);
+		MonetarySummaryStatistics summaryB = createSummaryDefault(DOLLAR);
 		MonetarySummaryStatistics result = summaryA.combine(summaryB);
 
 		assertEquals(BRAZILIAN_REAL, result.getCurrencyUnit());
@@ -84,6 +83,17 @@ public class ExchangeRateMonetarySummaryStatisticsTest {
 	private MonetarySummaryStatistics createSummary(CurrencyUnit currencyUnit) {
 		MonetarySummaryStatistics summary = new ExchangeRateMonetarySummaryStatistics(
 				currencyUnit, provider);
+		summary.accept(Money.of(10, currencyUnit));
+		summary.accept(Money.of(90, currencyUnit));
+		summary.accept(Money.of(110, currencyUnit));
+		return summary;
+	}
+
+	private MonetarySummaryStatistics createSummaryDefault(
+			CurrencyUnit currencyUnit) {
+		MonetarySummaryStatistics summary = new DefaultMonetarySummaryStatistics(
+				currencyUnit);
+
 		summary.accept(Money.of(10, currencyUnit));
 		summary.accept(Money.of(90, currencyUnit));
 		summary.accept(Money.of(110, currencyUnit));
