+++ src/test/java/org/javamoney/moneta/function/MonetaryFunctionsAggregatorTest.java
@@ -7,6 +7,8 @@ import static org.javamoney.moneta.function.MonetaryFunctions.min;
 import static org.javamoney.moneta.function.MonetaryFunctions.sum;
 import static org.javamoney.moneta.function.MonetaryFunctions.summarizingMonetary;
 import static org.javamoney.moneta.function.StreamFactory.BRAZILIAN_REAL;
+import static org.javamoney.moneta.function.StreamFactory.DOLLAR;
+import static org.javamoney.moneta.function.StreamFactory.EURO;
 import static org.javamoney.moneta.function.StreamFactory.currencies;
 import static org.javamoney.moneta.function.StreamFactory.currenciesToSummary;
 import static org.javamoney.moneta.function.StreamFactory.streamCurrencyDifferent;
@@ -21,23 +23,39 @@ import java.util.stream.Stream;
 import javax.money.CurrencyUnit;
 import javax.money.MonetaryAmount;
 import javax.money.MonetaryException;
+import javax.money.convert.ExchangeRateProvider;
 
+import org.javamoney.moneta.Money;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
 /**
- * Class to test the class ConversionOperators, but
+ * Class to test the class MonetaryFunctions, but
  * just the aggregator methods
  *
  * @author otaviojava
  */
 public class MonetaryFunctionsAggregatorTest {
 
+    private ExchangeRateProvider provider;
+
+    @Test
+    public void init() {
+        provider = new ExchangeRateProviderMock();
+    }
+
     @Test
     public void shouldSumCorrectly() {
         Stream<MonetaryAmount> stream = streamNormal();
         MonetaryAmount sum = stream.reduce(sum()).get();
         Assert.assertTrue(sum.getNumber().intValue() == 20);
+    }
+
+    @Test
+    public void shouldSumExchangeCorrectly() {
+        Stream<MonetaryAmount> stream = currencies();
+        MonetaryAmount sum = stream.reduce(sum(provider, DOLLAR)).get();
+        Assert.assertTrue(sum.getNumber().intValue() > 20);
     }
 
     @Test(expectedExceptions = NullPointerException.class)
@@ -59,6 +77,14 @@ public class MonetaryFunctionsAggregatorTest {
         Assert.assertTrue(min.getNumber().intValue() == 0);
     }
 
+    @Test
+    public void shouldMinExchangeCorretly() {
+        Stream<MonetaryAmount> stream = Stream.of(Money.of(7, EURO),
+                Money.of(9, BRAZILIAN_REAL), Money.of(8, DOLLAR));
+        MonetaryAmount min = stream.reduce(min(provider)).get();
+        Assert.assertEquals(Money.of(9, BRAZILIAN_REAL), min);
+    }
+
     @Test(expectedExceptions = NullPointerException.class)
     public void shouldMinWithNPEWhenAnElementIsNull() {
         Stream<MonetaryAmount> stream = streamNull();
@@ -77,6 +103,14 @@ public class MonetaryFunctionsAggregatorTest {
         MonetaryAmount max = stream.reduce(max()).get();
         Assert.assertTrue(max.getNumber().intValue() == 10);
     }
+
+    @Test
+    public void shouldMaxExchangeCorrectly() {
+        Stream<MonetaryAmount> stream = Stream.of(Money.of(7, EURO), Money.of(9, BRAZILIAN_REAL), Money.of(8, DOLLAR));
+        MonetaryAmount max = stream.reduce(max(provider)).get();
+        Assert.assertEquals(Money.of(7, EURO), max);
+    }
+
 
     @Test(expectedExceptions = NullPointerException.class)
     public void shouldMaxWithNPEWhenAnElementIsNull() {
@@ -117,11 +151,20 @@ public class MonetaryFunctionsAggregatorTest {
     }
 
     @Test
+    public void summarizingMonetaryExchangeTest() {
+        MonetarySummaryStatistics summary = currenciesToSummary().collect(
+                summarizingMonetary(BRAZILIAN_REAL,
+                        provider));
+
+        assertEquals(10L, summary.getCount());
+    }
+
+    @Test
     public void groupBySummarizingMonetaryTest() {
         GroupMonetarySummaryStatistics group = currenciesToSummary().collect(
                 groupBySummarizingMonetary());
         Map<CurrencyUnit, MonetarySummaryStatistics> mapSummary = group.get();
         assertEquals(mapSummary.keySet().size(), 3);
     }
-   
+
 }
