+++ src/test/java/org/javamoney/moneta/function/MonetaryFunctionsOrderTest.java
@@ -6,21 +6,29 @@ import static org.javamoney.moneta.function.MonetaryFunctions.sortNumber;
 import static org.javamoney.moneta.function.MonetaryFunctions.sortNumberDesc;
 import static org.javamoney.moneta.function.StreamFactory.BRAZILIAN_REAL;
 import static org.javamoney.moneta.function.StreamFactory.DOLLAR;
+import static org.javamoney.moneta.function.StreamFactory.EURO;
 import static org.javamoney.moneta.function.StreamFactory.currencies;
 
 import java.math.BigDecimal;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
 import javax.money.MonetaryAmount;
+import javax.money.convert.ExchangeRateProvider;
+
 import junit.framework.Assert;
 
+import org.javamoney.moneta.Money;
 import org.testng.annotations.Test;
 
 public class MonetaryFunctionsOrderTest {
 
-    //private ExchangeRateProvider provider;
+    private ExchangeRateProvider provider;
 
     @Test
     public void init() {
-        //provider = new ExchangeRateProviderMock();
+        provider = new ExchangeRateProviderMock();
     }
 
     @Test
@@ -58,5 +66,34 @@ public class MonetaryFunctionsOrderTest {
 
         Assert.assertEquals(BRAZILIAN_REAL, money.getCurrency());
         Assert.assertEquals(BigDecimal.ZERO, money.getNumber().numberValue(BigDecimal.class));
+    }
+
+    @Test
+    public void shouldExecuteValiableOrder() {
+
+        Stream<MonetaryAmount> stream = Stream.of(Money.of(7, EURO),
+                Money.of(9, BRAZILIAN_REAL), Money.of(8, DOLLAR));
+        List<MonetaryAmount> list = stream.sorted(
+                MonetaryFunctions.sortValuable(provider)).collect(
+                Collectors.toList());
+
+        Assert.assertEquals(Money.of(9, BRAZILIAN_REAL), list.get(0));
+        Assert.assertEquals(Money.of(8, DOLLAR), list.get(1));
+        Assert.assertEquals(Money.of(7, EURO), list.get(2));
+    }
+
+    @Test
+    public void shouldExecuteValiableOrderDesc() {
+
+        Stream<MonetaryAmount> stream = Stream.of(Money.of(7, EURO),
+                Money.of(9, BRAZILIAN_REAL), Money.of(8, DOLLAR));
+        List<MonetaryAmount> list = stream.sorted(
+                MonetaryFunctions.sortValuableDesc(provider)).collect(
+                Collectors.toList());
+
+        Assert.assertEquals(Money.of(7, EURO), list.get(0));
+        Assert.assertEquals(Money.of(8, DOLLAR), list.get(1));
+        Assert.assertEquals(Money.of(9, BRAZILIAN_REAL), list.get(2));
+
     }
 }
